Extract `getDynamicImport`
